Make the IDE tolerant to non-numeric team ids.
There didn't seem to be any reason that the IDE needed the ids to
be numeric, so trying to coerse them into being ints was the only
issue.

+++ include/auth/ldap.php
@@ -35,7 +35,7 @@ class LDAPAuth extends SecureTokenAuth
 			$groupNamePrefix = $config->getConfig("ldap.team.prefix");
 			if (stripos($group["cn"], $groupNamePrefix) === 0)
 			{
-				$teams[] = (int)substr($group["cn"], strlen($groupNamePrefix));
+				$teams[] = substr($group["cn"], strlen($groupNamePrefix));
 			}
 		}
